feat: migrate sources.js to sources.ts
